requirements.txt
Flask
Flask-Migrate

# environment
python-dotenv

# postgres dependencies
psycopg2-binary